requirements.txt
numpy
spectral
scipy
tqdm
visdom
seaborn
scikit-learn
scikit-image
matplotlib
joblib
einops
timm
linformer
batchgenerators
ml_collections
monai==1.3.0
nibabel==5.2.0
nni==3.0
numpy
opencv_python
Pillow
SimpleITK
Surface_Distance_Based_Measures==0.1
tifffile
torchtoolbox
unfoldNd
torchsummary
lightning
thop
torchsummaryX